requirements.txt
anthropic>=0.39.0
requests>=2.31.0
mcp>=1.21.1
python-dotenv>=1.2.1
